authorization with Gates

+++ app/Http/Controllers/ImageController.php
@@ -2,9 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\ImageRequest;
 use App\Models\Image;
 use Illuminate\Http\Request;
+use App\Http\Requests\ImageRequest;
+use Illuminate\Support\Facades\Gate;
 
 class ImageController extends Controller
 {
@@ -33,6 +34,7 @@ class ImageController extends Controller
 
     public function edit($image)
     {
+       Gate::authorize('update-image', $image);
         return view("image.edit", compact('image'));
     }
 
@@ -44,6 +46,9 @@ class ImageController extends Controller
 
     public function destroy(Image $image)
     {
+        if(Gate::denies('delete-image', $image)) {
+            abort(403, "Access denied");
+        }
         $image->delete();
         return to_route('images.index')->with('message', "Image has been removed successfully");
     }
